Finish first pass of cursor movement and scrolling
The user can now move a highlighted cursor around the displayed track
list in order to "select" different songs. When the user reaches the end
of what is displayed, the window will scroll. There is a slight bug here
when the user goes to the final track and then tries to scroll upward,
but it is only cosmetic in nature.

Also added is the `inchstr` function, which reads a string from the
display. This function is needed to move the cursor, and will also be
useful later for playing songs.

The existing function `truncate` was moved from within the `load_list`
function to the global scope, since it is useful to have in places
outside the `load_list` function.